Fix infyom.publish:templates
It looks like, when fixing issue #98, the PublishTemplateCommand was relocated, but the committer missed the use of `__DIR__` in the file, and didn't update the path to account for the additional directory depth, so `php artisan infyom.publish:templates` was only creating empty directories instead of actually publishing anything.  This commit fixes that.