fix: Swagger model property separator appearing on separate line
The current templates from InfyOmLabs/swagger-generator don't have an
EOL in the file whereas any normal unix editor will add this by default.

This trims trailing whitespace/line ends when building the list of
property annotations for a model and avoids the separator appearing on a
separate line for custom templates.